re-arranged login and register to listing setup
