+++ src/modules/user/routes.js
@@ -4,8 +4,7 @@ import { handleLogin } from "./controllers/controllerLogin.js";
 import { handleUpdate } from "./controllers/controllerUpdate.js";
 import { handleDelete } from "./controllers/controllerDelete.js";
 import validator from "express-validator";
-const { body, validationResult } = validator;
-import { verifyToken } from "../../utilities/jsonWebToken.js";
+const { body } = validator;
 import { handleRegister } from "./controllers/controllerRegister.js";
 
 export const usersRouter = express.Router();
@@ -13,63 +12,17 @@ export const usersRouter = express.Router();
 // POST /users
 usersRouter.post(
   "/",
-  // email must be actual email
   body("email").isEmail(),
-  // password must be at least 5 chars long
   body("password").isLength({ min: 5, max: 20 }),
-  async (req, res) => {
-    try {
-      // returns with errors if any
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return res.status(400).json({
-          message: `${errors.array()[0].msg} in ${errors.array()[0].param}${
-            errors.array()[1] ? " and " + errors.array()[1].param : ""
-          } field(s)`,
-        });
-      }
-      const data = await handleRegister(req);
-      if (data.status === 409) {
-        return res.status(409).json({ message: "User already exists" });
-      }
-      if (data.status === 400) {
-        return res.status(400).json({ message: "Bad image url" });
-      }
-      res.status(201).json(data);
-    } catch (error) {
-      // Send a 500 error if there was a problem with the insertion
-      console.error(error);
-      res.status(500).json({ ...error, message: "Internal server error" });
-    }
-  }
+  handleRegister
 );
 
 //  POST /users/login
 usersRouter.post(
-  "/login", // email must be actual email
+  "/login",
   body("email").isEmail(),
-  // password must be at least 5 chars long
   body("password").isLength({ min: 5, max: 20 }),
-  async (req, res) => {
-    try {
-      const errors = validationResult(req);
-      if (!errors.isEmpty()) {
-        return res.status(400).json({
-          message: `${errors.array()[0].msg} in ${errors.array()[0].param}${
-            errors.array()[1] ? " and " + errors.array()[1].param : ""
-          } field(s)`,
-        });
-      }
-
-      const { status, data } = await handleLogin(req);
-
-      console.log(await verifyToken(data.token));
-      res.status(status).json(data);
-    } catch (error) {
-      console.log(error);
-      res.status(500).json({ ...error, message: "Internal server error." });
-    }
-  }
+  handleLogin
 );
 
 // GET /users
